requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.6.1
Markdown==3.3.4
numpy==1.21.1
packaging==21.0
pandas==1.3.0
Pint==0.17
plotly==5.1.0
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.2.4
tenacity==8.0.1
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
